Add files via upload
This is an assignment for appointment bookings.  This is a project developed using native php, with mariadb as the database. The project has a login, and receptionist and doctors pages based on the logged in  user. To test, install xamp which has mariadb  and inport the .sql file. The db has sampe users to test with. Receiptionist can onboard a patient, book doctors appointment for the patient, update appointment details, doctor can see bookings, and history of the patient. Please note the system is not developed to concluion. Mostl it captured the all design of the booking process from onboard to being listing in doctors appointment list. The pages depend on role given to looged in user. Note also the security aspect has not been implemented to the standard due to time constrain.